requirements.txt
autopep8==1.6.0
click==8.0.4
dnslib==0.9.19
dnspython==2.2.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-WTF==1.0.0
importlib-metadata==4.12.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
python-ldap==3.4.0
pytz==2021.3
six==1.16.0
toml==0.10.2
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.1
